Allow more flexible output options
This commit is the start of a refactoring which will
close #17, by allowing more flexible output generation.

I've moved the instance-fetching code to a common routine,
and updated the existing subcommands to use it, with no
changes.

The next step will be to allow changing the output of
each subcommand, while leaving the default(s) as-is.
